+++ src/main/java/se331/lab/service/EventService.java
@@ -6,6 +6,6 @@ import se331.lab.entity.Event;
 public interface EventService {
     Integer getEventSize();
     Page<Event> getEvents(Integer pageSize, Integer page);
-    Event getEvent(Integer id);
+    Event getEvent(Long id);
     Event save(Event event);
 }
